Rename Share component
